Re-organize map impls
