+++ runtime/CMakeLists.txt
@@ -66,15 +66,15 @@ set(sources
   src/ffi.cpp
   src/bpf_helper.cpp
 
-  src/bpf_map/array_map.cpp
-  src/bpf_map/hash_map.cpp
-  src/bpf_map/ringbuf_map.cpp
-  src/bpf_map/perf_event_array_map.cpp
-  src/bpf_map/per_cpu_array_map.cpp
-  src/bpf_map/per_cpu_hash_map.cpp
-  src/bpf_map/array_map_kernel_user.cpp
-  src/bpf_map/hash_map_kernel_user.cpp
-  src/bpf_map/percpu_array_map_kernel_user.cpp
+  src/bpf_map/userspace/array_map.cpp
+  src/bpf_map/userspace/hash_map.cpp
+  src/bpf_map/userspace/ringbuf_map.cpp
+  src/bpf_map/userspace/perf_event_array_map.cpp
+  src/bpf_map/userspace/per_cpu_array_map.cpp
+  src/bpf_map/userspace/per_cpu_hash_map.cpp
+  src/bpf_map/shared/array_map_kernel_user.cpp
+  src/bpf_map/shared/hash_map_kernel_user.cpp
+  src/bpf_map/shared/percpu_array_map_kernel_user.cpp
 )
 
 # list(APPEND sources
