+++ runtime/src/bpf_map/shared/hash_map_kernel_user.cpp
@@ -1,7 +1,5 @@
 #include "spdlog/spdlog.h"
-#include <bpf_map/hash_map_kernel_user.hpp>
-#include <algorithm>
-#include <functional>
+#include <bpf_map/shared/hash_map_kernel_user.hpp>
 #include <unistd.h>
 #include <bpf/bpf.h>
 #include <bpf/libbpf.h>
