+++ runtime/unit-test/maps/test_per_cpu_array.cpp
@@ -1,7 +1,7 @@
 #include "catch2/catch_message.hpp"
 #include <boost/interprocess/creation_tags.hpp>
 #include <boost/interprocess/interprocess_fwd.hpp>
-#include <bpf_map/per_cpu_array_map.hpp>
+#include <bpf_map/userspace/per_cpu_array_map.hpp>
 #include <catch2/catch_test_macros.hpp>
 #include <cstdint>
 #include <sched.h>
